Organização das rotas do backend e ajuste das rotas no frontend
