+++ src/Backend/routes/AuthRoutes.js
@@ -7,3 +7,4 @@ router.post('/signup', authController.register)
 router.post('/login', authController.login)
 
 module.exports = router
+
