+++ src/Backend/routes/EcoPontoRoutes.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+const ecoPonto = require(`../controllers/EcoPontosController`)
+
+router.post('/cadastro/ecopontos', ecoPonto.registerEcoPonto)
+router.get('/pesquisa/:cep/:raio', ecoPonto.findNearbyEcoPonto)
+
+module.exports = router
