+++ src/Frontend/ecorota/src/pages/Sign.js
@@ -135,7 +135,7 @@ const Login = () =>{
         return
       }
 
-      const response = await fetch (`https://projeto14-fork-728609929656.southamerica-east1.run.app/auth/${endpoint}`,{
+      const response = await fetch (`${process.env.REACT_APP_ECOROTA_API_HOST}/auth/${endpoint}`,{
         method:'POST',
         headers: { 'Content-Type' : 'application/json'},
         body: JSON.stringify({ email, password, name })
